feat(doctrine): DoctrineSystem — per-org doctrine loop wired into the tick (Unit 4)
The 28th engine system (position 14.7, after OODA/FactionInfluence). Each tick per
organization: decay tags (0.55%/tick) -> accrue theoretical labour
(study_allocation x cadre_level) -> bootstrap free root -> greedy auto-acquire
cheapest affordable unlocked node -> fire any reachable trap whose condition holds.
Deterministic (no RNG). Registered in _DEFAULT_SYSTEMS + CONSEQUENCE_SYSTEMS.
evaluate_trap_condition widened to Mapping[DoctrineTag, float] for the accumulator.

Events + behavioural feedback into bifurcation/consciousness land in Unit 6.
Verified: 7 system + 40 mechanics TDD cases, qa:regression 5/5 byte-identical
(scenarios carry no org nodes -> no-op), 14 engine partition tests.

+++ src/babylon/domain/doctrine/mechanics.py
@@ -35,7 +35,9 @@ class DoctrineExpressionError(ValueError):
 
 
 #: Comparison operators the DSL supports, longest-token-first for the tokenizer.
-_COMPARISONS: dict[str, Callable[[int, int], bool]] = {
+#: Left operand is a tag total (``float`` accumulator or ``int``); right is the
+#: literal ``int`` threshold from the condition string.
+_COMPARISONS: dict[str, Callable[[float, int], bool]] = {
     "<=": lambda a, b: a <= b,
     ">=": lambda a, b: a >= b,
     "==": lambda a, b: a == b,
@@ -83,7 +85,7 @@ class _Parser:
     remaining input.
     """
 
-    def __init__(self, tokens: list[str], tags: Mapping[DoctrineTag, int], expr: str) -> None:
+    def __init__(self, tokens: list[str], tags: Mapping[DoctrineTag, float], expr: str) -> None:
         self._tokens = tokens
         self._i = 0
         self._tags = tags
@@ -156,7 +158,7 @@ class _Parser:
         return op(self._tags.get(tag, 0), threshold)
 
 
-def evaluate_trap_condition(condition: str, tags: Mapping[DoctrineTag, int]) -> bool:
+def evaluate_trap_condition(condition: str, tags: Mapping[DoctrineTag, float]) -> bool:
     """Return whether ``condition`` holds against the current doctrine ``tags``.
 
     :param condition: A ``trap_condition`` expression in the DSL documented in
